Refactored ServersView class into several classes
This change refactors the ServersView class into
several classes to make it easier to read and
understand. The ColumnFactory interface and
implementations, and the Status class, were moved
into their own files. The ServersView object
was renamed to TableData as it represents a
generic DTO object that is comprised of a
list of columns and the corresponding metric
data for those columns. The logic for creating
the TableData objects was moved into the new
TableDataFactory class.

Closes #6341
