+++ server/monitor/src/main/java/org/apache/accumulo/monitor/next/views/RatioColumnFactory.java
@@ -1,0 +1,67 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.accumulo.monitor.next.views;
+
+import java.util.List;
+import java.util.Map;
+
+import org.apache.accumulo.core.client.admin.servers.ServerId;
+import org.apache.accumulo.core.metrics.Metric;
+import org.apache.accumulo.core.metrics.flatbuffers.FMetric;
+import org.apache.accumulo.core.process.thrift.MetricResponse;
+import org.apache.accumulo.monitor.next.views.TableData.Column;
+
+public class RatioColumnFactory implements ColumnFactory {
+
+  private final Column column;
+  private final Metric numerator;
+  private final Metric denominator;
+
+  RatioColumnFactory(String label, String description, Metric numerator, Metric denominator) {
+    this.column = new Column(numerator.getName() + "/" + denominator.getName(), label, description,
+        Metric.MonitorCssClass.PERCENT.getCssClass());
+    this.numerator = numerator;
+    this.denominator = denominator;
+  }
+
+  @Override
+  public Column getColumn() {
+    return column;
+  }
+
+  @Override
+  public Object getRowData(ServerId sid, MetricResponse mr,
+      Map<String,List<FMetric>> serverMetrics) {
+    var n = serverMetrics.get(numerator.getName());
+    var d = serverMetrics.get(denominator.getName());
+
+    if (n == null || d == null) {
+      return null;
+    }
+
+    var numeratorSum = sum(n).doubleValue();
+    var denominatorSum = sum(n).doubleValue();
+
+    if (denominatorSum == 0) {
+      return null;
+    }
+
+    return numeratorSum / denominatorSum;
+  }
+}
